Several minor fixes found testing T41-2.

+++ T41EEE/Encoders.cpp
@@ -544,8 +544,8 @@ uint32_t SetTransmitDelay() {
 *****/
 void EncoderFineTune() {
   char result;
-  unsigned long MS_temp;   // HMB
-  unsigned long FT_delay;  // HMB
+  uint32_t MS_temp;   // HMB
+  uint32_t FT_delay;  // HMB
 
   result = fineTuneEncoder.process();  // Read the encoder
   if (result == 0) {                   // Nothing read
@@ -580,7 +580,7 @@ void EncoderFineTune() {
   }
 
   NCOFreq = NCOFreq + ConfigData.fineTuneStep * fineTuneEncoderMove;  // Increment NCOFreq per encoder movement.
-  centerTuneFlag = 1;                                                 // This is used in Process.cpp.  Greg KF5N May 16, 2024
+  centerTuneFlag = 1;                                                 // This is used in ReceiveDSP.cpp.  Greg KF5N May 16, 2024
   // ============  AFP 10-28-22
   if (ConfigData.activeVFO == VFO_A) {
     ConfigData.currentFreqA = ConfigData.centerFreq + NCOFreq;  //AFP 10-05-22
